Stash CPU policy related extra_spec in system_metadata
This is a temporary measure before we have the flavor properly saved
with the instance (see: https://review.openstack.org/#/c/126363/)

This is really only needed for resize claims to work as expected. as all
the info related to CPU pinning will be saved in the instance_extra
table on boot.

Change-Id: I90d3fc5e98fe3079a493c4cb169e830d99b01f62
Blueprint: virt-driver-cpu-pinning