Make get_best_cpu_topology consider NUMA topology
CPU pinning logic may expose a topology per NUMA cell that we should
consider when calculating the final topology for the instance. This
patch makes sure that if threads are exposed in a NUMA cell - they will
be exposed in the final cpu_topology that instance gets

CPU pinning logic at the moment exposes thread and core information. In
case there is unlikely a different topology exposed per cell, it's
safest to expose the least amount of threads to the whole instance.

As we still don't call any of the CPU pinning code when booting an
instance - these code paths won't get hit yet.

Change-Id: Ibe6c3b331253b36fc9e1338c6f0f83bc9094e3dc
Blueprint: virt-driver-cpu-pinning