libvirt: un-cruft _get_guest_numa_config
Several issues get cleaned up with this patch - all of which will make
adding CPU pinning into the mix a lot cleaner.

1) Unecessary context argument to the method is dropped.

2) Secondly - we pass in the InstanceNUMATopology object (or None)
instead of passing in the whole instance and then getting the topology
out - this is useful as we will need instance NUMA for generating topology
as well, and this avoids repetitive code in a lot of places as we now grab
it in the top level method, and just pass it in where needed.

Change-Id: I829465d1612f42e7af92927584b31ca6a2349d5a

Conflicts:
	nova/virt/libvirt/driver.py