libvirt: Fix NUMA memnode assignments to host cells
This fix

  https://review.openstack.org/138128

breaks assignments for NUMA topologies where host cell ID
that the instance cell is pinned to is not in increasing order. The
reason for this is that the main loop of the method that will go through
the host cells (always in ascending order), and find instance cells that
match (may not be in ascending order depending on how the fitting logic
and usage of the host) but we simply append memnodes so they may end up
in the wrong order when we hit the normalization loop.

We fix this by "pre-allocating" the memnodes list and making sure to add
the matching memnode to the same position as the guest config cell
before hitting the normalization loop.

Conflicts:
	nova/tests/virt/libvirt/test_driver.py
	nova/virt/libvirt/driver.py

Closes-bug: #1397381
Change-Id: I91055dc705f2ef54c830e641f9cf9ba0683eb313