Add method for getting the CPU pinning constraint
We add a method that when passed flavor with extra spec containing
hw:cpu_policy key, and an optional NUMA topology instance,
calculates the CPU pinning constraint for the instance (or None if
the instance is not be pinned), and set the cpu_pinning attribute of the
passed NUMA topology (or a constructed one).

In case admin doesn't specify a shared pinning policy on the flavor,
users are free to override it with image metadata hw_cpu_policy,
otherwise, attempts to override it will result in an error.

This will be done at boot time and then persisted to be used later
in scheduling.

This patch concludes

Blueprint: virt-driver-cpu-pinning

Change-Id: I17b4eab368fca6ddb7e000f9c24368bbde18534d