Ironic: Adds confidrive support
This patch is adding configdrive v2 support for the Ironic drive. As
attaching a volume in a baremetal instance is not as simple as VMs and
also hardware depedent, this patch will base64 + gzip the generated
configdrive and update the Ironic node with it, from there Ironic can
create an extra parition on the local disk of the machine and copy the
configdrive onto it.

Implements: blueprint use-configdrive-with-ironic
Change-Id: Ibc745e9a63bc3d5d68217020b3d591f71c3f8a96

Conflicts:
	nova/tests/unit/virt/ironic/test_driver.py